Refactor ConstructiveSolidGeometry.jl into smaller files containing specific topics such as types or geometric operators. Otherwise, code remains unchanged.
